+++ src/palm-react/hooks/page/explore/useExploreRecommendChat.ts
@@ -1,5 +1,9 @@
+import { RECOMMENDED_CHANNELS_URL } from 'palm-core/consts/url'
+import { UTIL } from 'palm-core/libs'
+import fetchWithTimeout from 'palm-core/libs/fetchWithTimeout'
 import { FbChannel } from 'palm-core/types'
 import useFsChannels from 'palm-react/hooks/firestore/useFsChannels'
+import { useEffect, useState } from 'react'
 
 export type UseExploreRecommendChatReturn = {
   fsChannelList: FbChannel[]
@@ -7,7 +11,37 @@ export type UseExploreRecommendChatReturn = {
 }
 
 const useExploreRecommendChat = (): UseExploreRecommendChatReturn => {
-  const { isFetching, channels } = useFsChannels()
+  const [recommendedChannelUrls, setRecommendedChannelUrls] = useState<
+    string[] | undefined
+  >(undefined)
+
+  const { isFetching, channels } = useFsChannels({
+    recommendedChannelUrls,
+  })
+
+  const fetchRecommendedChatUrls = async (): Promise<void> => {
+    try {
+      const ret = await fetchWithTimeout(RECOMMENDED_CHANNELS_URL, 5000)
+      if (ret.ok) {
+        const urls = (await ret.json()) as {
+          testnet: string[]
+          mainnet: string[]
+        }
+        setRecommendedChannelUrls(
+          UTIL.isMainnet() ? urls.mainnet : urls.testnet
+        )
+      } else {
+        setRecommendedChannelUrls([])
+      }
+    } catch (e) {
+      setRecommendedChannelUrls([])
+      console.error(e)
+    }
+  }
+
+  useEffect(() => {
+    fetchRecommendedChatUrls()
+  }, [])
 
   return {
     fsChannelList: channels,
